+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @glissade/eslint-plugin
 
+## 0.51.0-pre.1
+
 ## 0.51.0-pre.0
 
 ## 0.50.0